+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "app-vuejs-demo",
-  "title": "Marvel Admin 🎉 ! ",
+  "title": "Marvel Admin ! ",
   "version": "1.0.0",
   "private": true,
   "scripts": {
@@ -9,7 +9,8 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "@mdi/font": "^4.7.95", 
+    "@mdi/font": "^4.7.95",
+    "@vuejs-community/vue-filter-date-format": "^1.6.3",
     "axios": "^0.21.1",
     "core-js": "^3.6.5",
     "normalize.css": "^8.0.1",
