+++ src/main.js
@@ -12,14 +12,16 @@ import './registerServiceWorker'
 import i18n from './locate/index'; // Internationalization
 import '@/scss/style.scss'
 import PerfectScrollbar from 'vue2-perfect-scrollbar'
-import 'vue2-perfect-scrollbar/dist/vue2-perfect-scrollbar.css'   
+import 'vue2-perfect-scrollbar/dist/vue2-perfect-scrollbar.css'    
+import VueFilterDateFormat from '@vuejs-community/vue-filter-date-format';
 Vue.config.productionTip = false
 
+Vue.use(VueFilterDateFormat);
 new Vue({
   router,
   store,
   vuetify,
   PerfectScrollbar,
-  i18n, 
+  i18n,  
   render: h => h(App)
 }).$mount('#app')
