+++ src/views/components/Center.vue
@@ -19,3 +19,4 @@ export default {
   data: () => ({}),
 };
 </script>
+© 2021 GitHub, Inc.
